Disable C4251 with MSVC
This is a warning about:

'identifier': class 'type' needs to have dll-interface to be used by
clients of class 'type2'

which in this case refers to the use of std::unique_ptr to implement
the pimpl idiom in the rlottie::Animation class.

In this case the pimpl pointer is not accessible by clients of the class
so we can simply disable the warning.